+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `GetBitmapFrameByWidthAndDpi` to round `dpi` frame value.
 - Update `Width` to `Math.Round` to improve order by `Width`.
 - Update `GetBitmapFrameByWidthAndDpi` to public.
+- Update `GetBitmapFrameByWidthAndDpi` to return last `Width`.
+- Update `UriToBitmapFrame` to use `GetBitmapFrameByWidthAndDpi` with `int.MaxValue`.
 ### Example
 - Add `Cube-Grey-Light.tiff` and `Cube-Grey-Dark.tiff` in `AppTheme`.
 ### Tests